requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==2.9.2
Django==3.1.1
django-cuser==2017.3.16
django-restframework==0.0.1
djangorestframework==3.11.1
drf-access-policy==0.7.0
dry-rest-permissions==0.1.10
environ==1.0
Faker==0.9.1
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
iniconfig==1.0.1
install==1.3.3
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mixer==6.1.3
mock==4.0.2
more-itertools==8.5.0
openapi-codec==1.3.2
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.0.1
pytest-cov==2.10.1
pytest-django==3.9.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
zipp==3.1.0
